step-1 and 2 separated and connected to store
