+++ src/components/App.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import {BrowserRouter as Router, Route} from 'react-router-dom';
+
 import Navigation from './Navigation';
 import LandingPage from './Landing';
 import SignUpPage from './SignUp';
@@ -11,7 +12,8 @@ import AccountPage from './Account';
 import * as routes from '../constants/routes';
 import withAuthentication from './withAuthentication';
 
-import ProfilePage from '../pages/profile/create';
+import Step1 from '../pages/profile/step-1';
+import Step2 from '../pages/profile/step-2';
 
 
 const App = () =>
@@ -25,7 +27,8 @@ const App = () =>
       <Route exact path={routes.PASSWORD_FORGET} component={() => <PasswordForgetPage/>}/>
       <Route exact path={routes.HOME} component={() => <HomePage/>}/>
       <Route exact path={routes.ACCOUNT} component={() => <AccountPage/>}/>
-      <Route exact path={routes.CREATE} component={() => <ProfilePage/>}/>
+      <Route exact path={routes.STEP_1} component={() => <Step1/>}/>
+      <Route exact path={routes.STEP_2} component={() => <Step2/>}/>
     </div>
   </Router>;
