+++ src/components/Checkbox.js
@@ -2,9 +2,9 @@ import React from 'react';
 
 export default (props) => {
   return (
-    <div key={props.index} className="checkbox-wrapper">
-      <label className="checkbox" htmlFor={props.name}>
-        <input type="checkbox" name={props.name} checked={props.checked} onClick={props.onClick} value={props.name}/>{props.name}
+    <div key={props.index} className={`checkbox-wrapper${props.type ? " slider" : ""}`}>
+      <label className="checkbox">
+        <input type="checkbox" checked={props.checked} onClick={props.onClick} value={props.name}/><span>{props.name}</span>
       </label>
     </div>
   )
