+++ src/constants/routes.js
@@ -4,5 +4,6 @@ export const LANDING = '/';
 export const HOME = '/home';
 export const ACCOUNT = '/account';
 export const PASSWORD_FORGET = '/pw_forget';
-export const CREATE = '/profile/create';
+export const STEP_1 = '/profile/step-1';
+export const STEP_2 = '/profile/step-2';
