+++ src/index.js
@@ -1,10 +1,17 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import {Provider} from 'react-redux';
 
 import App from './components/App';
+import store from './store';
+
 import './styles/global.css';
 
 import registerServiceWorker from './registerServiceWorker';
 
-ReactDOM.render(<App />, document.getElementById('root'));
+ReactDOM.render(
+  <Provider store={store}>
+    <App/>
+  </Provider>
+  , document.getElementById('root'));
 registerServiceWorker();
