+++ src/reducers.js
@@ -1,0 +1,8 @@
+import { combineReducers } from 'redux';
+
+
+const rootReducer = combineReducers({
+
+});
+
+export default rootReducer;
